+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 FEATURES:
 
+- **New Resource:** `netapp-ontap_fpolicy_policy` ([#623](https://github.com/NetApp/terraform-provider-netapp-ontap/issues/623))
 - **New Resource:** `netapp-ontap_fpolicy_event` ([#624](https://github.com/NetApp/terraform-provider-netapp-ontap/issues/624))
 - **New Data Source:** `netapp-ontap_network_ip_service_policy` and `netapp-ontap_network_ip_service_policies`([#621](https://github.com/NetApp/terraform-provider-netapp-ontap/issues/621))
 - **New Resource:** `netapp-ontap_network_ip_service_policy` ([#621](https://github.com/NetApp/terraform-provider-netapp-ontap/issues/621))